modify et delet fonctionnel

+++ backend/controllers/bookController.js
@@ -1,3 +1,4 @@
+const fs = require('fs');
 const Book = require('../models/bookModel');
 
 exports.getAllBooks = (req, res, next) => {
@@ -17,6 +18,7 @@ exports.getBookById = (req, res, next) => {
     .catch(error => res.status(400).json({ error }));
 };
 
+
 exports.createBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book);
   delete bookObject._id;
@@ -32,14 +34,51 @@ exports.createBook = (req, res, next) => {
   .catch(error => { res.status(400).json( { error })})
 };
 
-exports.updateBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Livre mis à jour avec succès' }))
-    .catch(error => res.status(400).json({ error }));
+exports.modifyBook = (req, res, next) => {
+  const bookObject = req.file ? {
+      ...JSON.parse(req.body.book),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body };
+  
+  delete bookObject._userId;
+  
+  Book.findOne({_id: req.params.id})
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'Livre modifié !' }))
+          .catch(error => res.status(401).json({ error }));
+      }
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Livre supprimé avec succès' }))
+  Book.findOne({ _id: req.params.id })
+    .then(book => {
+      if (!book) {
+        return res.status(404).json({ message: 'Livre non trouvé' });
+      }
+      if (book.userId !== req.auth.userId) {
+        return res.status(401).json({ message: 'Non autorisé à supprimer ce livre' });
+      }
+      const filename = book.imageUrl.split('/images/')[1];
+
+      fs.unlink(`images/${filename}`, () => {
+        Book.deleteOne({ _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'Livre supprimé avec succès' }))
+          .catch(error => res.status(400).json({ error }));
+      });
+    })
+    .catch(error => res.status(500).json({ error }));
+};
+
+exports.getBestRatedBooks = (req, res, next) => {
+  Book.find().sort({ averageRating: -1 }).limit(3)
+    .then(books => res.status(200).json(books))
     .catch(error => res.status(400).json({ error }));
 };
